split into lnurl.go and server.go
